feat: code optimization
ALMS-127
