+++ library/src/main/java/software/amazon/location/auth/data/response/Credentials.kt
@@ -1,0 +1,10 @@
+package software.amazon.location.auth.data.response
+
+import com.google.gson.annotations.SerializedName
+
+data class Credentials(
+    @SerializedName("AccessKeyId") val accessKeyId: String,
+    @SerializedName("Expiration") val expiration: Double,
+    @SerializedName("SecretKey") val secretKey: String,
+    @SerializedName("SessionToken") val sessionToken: String
+)
